(PC-27682)[PRO] fix: move venue label to settings page
