+++ pro/src/pages/VenueEdition/setInitialFormValues.ts
@@ -26,6 +26,5 @@ export const setInitialFormValues = (
     isAccessibilityAppliedOnAllOffers: false,
     phoneNumber: venue.contact?.phoneNumber || '',
     webSite: venue.contact?.website || '',
-    venueLabel: venue.venueLabelId?.toString() || '',
   }
 }
